🐭 Put the ShapePage store in the state initializer
